+++ examples/get_it_signals/lib/main.dart
@@ -145,7 +145,7 @@ class DarkModeToggle extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Watch((_) {
+    return Watch((_, __) {
       final isDark = brightness() == Brightness.dark;
       return IconButton(
         onPressed: () {
@@ -192,7 +192,7 @@ class _HomeScreenState extends State<HomeScreen> {
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
             const Text('Count value:'),
-            Watch((context) {
+            Watch((context, _) {
               return Text(
                 '$counter',
                 style: Theme.of(context).textTheme.headlineMedium,
@@ -303,7 +303,7 @@ class ProfileScreen extends StatelessWidget {
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
             const Text('Profile'),
-            Watch((context) {
+            Watch((context, _) {
               return Text(
                 getIt.get<Auth>().currentUserName(),
                 style: Theme.of(context).textTheme.headlineMedium,