+++ examples/node_based_editor/lib/nodes/datetime.dart
@@ -24,7 +24,7 @@ class DateTimeNode extends ValueNode<DateTime> {
   @override
   Widget build() {
     if (output is Signal<DateTime>) {
-      return Watch.builder(builder: (context) {
+      return Watch.builder(builder: (context, _) {
         return InkWell(
           onTap: () {
             final now = DateTime.now();